+++ q_station_info.h
@@ -6,7 +6,7 @@
 
 class QStationInfo: public QObject {
     Q_OBJECT
-    Q_PROPERTY(unsigned int audioKbps READ audioKbps NOTIFY audioKbpsChanged)
+    Q_PROPERTY(size_t audioKbps READ audioKbps NOTIFY audioKbpsChanged)
     Q_PROPERTY(QString stationName READ stationName NOTIFY stationNameChanged)
 
     size_t audio_kbps_;
@@ -15,7 +15,7 @@ class QStationInfo: public QObject {
 public:
     explicit QStationInfo(QObject* parent = 0);
 
-    unsigned int audioKbps() const;
+    size_t audioKbps() const;
     void setAudioKbps(size_t audio_kbps);
     QString stationName() const;
     void setStationName(std::string station_name);
